Move some assertion logic from problem.py to asserts.py
